allow roc warnings in build script and suppress rust warnings

+++ build.sh
@@ -1,5 +1,5 @@
 #!/bin/bash
-set -euo pipefail
+set -uo pipefail
 
 if [ -z "$1" ]; then
     echo "Error: No file name provided."
@@ -9,6 +9,12 @@ fi
 
 roc build $1 --lib --output libapp.so --target linux-x64
 
+roc_exit_code=$?
+# Exit code 2 means there were warnings but no errors
+if [[ $roc_exit_code -ne 0 && $roc_exit_code -ne 2 ]]; then
+    exit $roc_exit_code
+fi
+
 # We need to include libapp.so.1 in the final zip archive
 mv libapp.so.1.0 libapp.so.1
 
@@ -16,7 +22,11 @@ mv libapp.so.1.0 libapp.so.1
 cp libapp.so.1 platform/libapp.so
 
 cd platform
-RUSTFLAGS="-C link-args=-rdynamic" cargo build --release
+RUSTFLAGS="-C link-args=-rdynamic" cargo build --release --target x86_64-unknown-linux-gnu
+cargo_exit_code=$?
+if [ $cargo_exit_code -ne 0 ]; then
+    exit $cargo_exit_code
+fi
 rm libapp.so
 cp target/release/host ../bootstrap
